Factor out patch_additional_properties
